refactor: extract shared assessment-write helpers into _assessment_group
