+++ src/routes/_assessment_group.py
@@ -1,0 +1,130 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2026 Savoir-faire Linux, Inc.
+# SPDX-License-Identifier: GPL-3.0-only
+
+"""Shared helpers for assessment write endpoints."""
+
+from datetime import datetime
+from uuid import UUID
+
+from ..extensions import db
+from ..models import Assessment as DBAssessment, Finding, Package
+from ..models.assessment import STATUS_TO_SIMPLIFIED
+
+
+def resolve_package(pkg_string_id: str) -> "Package | None":
+    """Look up an existing Package for 'name@version::supplier'.
+
+    Returns ``None`` when no matching package exists. Writing an assessment must
+    never create a package, so callers block the write when this returns
+    ``None``. Matching is on name + version + supplier (with the same supplier
+    normalization used by :meth:`Package.find_or_create`).
+    """
+    return Package.get_by_string_id(pkg_string_id)
+
+
+def find_valid_finding(package_id: UUID, vuln_id: str, variant_id: UUID) -> "Finding | None":
+    """Return the finding when it was actually observed for the variant.
+
+    Assessment writes may target active or historical package versions, but
+    they must never invent a package/vulnerability/variant relationship that
+    was not produced by a scan.
+    """
+    from ..models.observation import Observation
+    from ..models.scan import Scan
+
+    return db.session.execute(
+        db.select(Finding)
+        .join(Observation, Observation.finding_id == Finding.id)
+        .join(Scan, Scan.id == Observation.scan_id)
+        .where(
+            Finding.package_id == package_id,
+            Finding.vulnerability_id == vuln_id.upper(),
+            Scan.variant_id == variant_id,
+        )
+        .distinct()
+    ).scalar_one_or_none()
+
+
+def validate_assessment_findings(
+    packages: list[Package], vuln_id: str, variant_id: UUID
+) -> "tuple[dict[UUID, Finding], list[str]]":
+    findings: dict[UUID, Finding] = {}
+    invalid: list[str] = []
+    for package in packages:
+        finding = find_valid_finding(package.id, vuln_id, variant_id)
+        if finding is None:
+            invalid.append(package.string_id)
+        else:
+            findings[package.id] = finding
+    return findings, invalid
+
+
+def create_assessment_record(
+    assessment: "DBAssessment",
+    finding_id: UUID,
+    variant_id: UUID | None,
+    timestamp: datetime | None = None,
+    origin: str = "custom",
+) -> "DBAssessment":
+    """Create a single DBAssessment row from a validated DTO.
+
+    Shared between ``add_assessment`` (single) and ``add_assessments_batch``.
+    """
+    return DBAssessment.create(
+        status=assessment.status or "",
+        simplified_status=STATUS_TO_SIMPLIFIED.get(assessment.status or "", "Pending Assessment"),
+        finding_id=finding_id,
+        variant_id=variant_id,
+        origin=origin,
+        status_notes=assessment.status_notes,
+        justification=assessment.justification,
+        impact_statement=assessment.impact_statement,
+        workaround=getattr(assessment, "workaround", None),
+        responses=list(assessment.responses) if assessment.responses else [],
+        commit=True,
+        timestamp=timestamp,
+    )
+
+
+def payload_to_assessment(data: dict) -> "tuple[DBAssessment | dict[str, str], int]":
+    """
+    Take an object in input and try to convert it to an Assessment DTO.
+    Return either (Assessment, 200) or (error_dict, http_code).
+    """
+    if "packages" not in data or not isinstance(data["packages"], list) or len(data["packages"]) < 1:
+        return {"error": "Invalid request data"}, 400
+
+    assessment = DBAssessment.new_dto(data["vuln_id"], data["packages"])
+
+    if "status" not in data or not isinstance(data["status"], str):
+        return {"error": "Invalid request data"}, 400
+
+    if assessment.set_status(data["status"]) is False:
+        return {"error": "Invalid status"}, 400
+
+    if "status_notes" in data and isinstance(data["status_notes"], str):
+        assessment.set_status_notes(data["status_notes"], False)
+
+    if "justification" in data and isinstance(data["justification"], str):
+        if not assessment.set_justification(data["justification"]):
+            return {"error": "Invalid justification"}, 400
+    elif assessment.is_justification_required():
+        return {"error": "Justification required"}, 400
+
+    if "impact_statement" in data and isinstance(data["impact_statement"], str):
+        assessment.set_not_affected_reason(data["impact_statement"], False)
+
+    if "workaround" in data and isinstance(data["workaround"], str):
+        assessment.set_workaround(data["workaround"])
+
+    if "timestamp" in data and isinstance(data["timestamp"], str):
+        try:
+            assessment.timestamp = datetime.fromisoformat(data["timestamp"])
+        except (ValueError, TypeError):
+            pass
+    if "responses" in data and isinstance(data["responses"], list):
+        for response in data["responses"]:
+            assessment.add_response(response)
+    return assessment, 200
